+++ packages/modules/src/firebase.ts
@@ -7,6 +7,9 @@ import serveModulesApi from './firebase/serve-modules-api';
 
 const { httpsFunctionOptions } = config.get();
 
-export const modules = onRequest(httpsFunctionOptions, (req, res) => {
+export const modules = onRequest({
+  concurrency: 6,
+  ...httpsFunctionOptions,
+}, (req, res) => {
   serveModulesApi(req, res);
 });