+++ packages/storefront/astro.config.mjs
@@ -8,12 +8,124 @@ import image from '@astrojs/image';
 import partytown from '@astrojs/partytown';
 import prefetch from '@astrojs/prefetch';
 import sitemap from '@astrojs/sitemap';
+import UnoCSS from 'unocss/astro';
+import { VitePWA } from 'vite-plugin-pwa';
 import getConfig from './storefront.config.mjs';
 
 dotenv.config();
 
-// https://astro.build/config
-export default defineConfig({
+const { domain, primaryColor, settings } = getConfig();
+
+const _vitePWAOptions = {
+  manifest: {
+    name: settings.name || 'My Shop',
+    short_name: settings.short_name || settings.name || 'MyShop',
+    description: settings.description || 'My PWA Shop',
+    background_color: settings.bg_color || '#ffffff',
+    theme_color: primaryColor,
+    crossorigin: 'use-credentials',
+    icons: [{
+      src: settings.icon || '/img/icon.png',
+      sizes: '192x192',
+      type: 'image/png',
+    }, {
+      src: settings.large_icon || '/img/large-icon.png',
+      sizes: '512x512',
+      type: 'image/png',
+    }, {
+      src: settings.large_icon || '/img/large-icon.png',
+      sizes: '512x512',
+      type: 'image/png',
+      purpose: 'any maskable',
+    }],
+  },
+  registerType: 'autoUpdate',
+  workbox: {
+    navigateFallback: null,
+    globDirectory: 'dist/client',
+    globPatterns: [
+      '**/!(cms*|admin*).{js,css}',
+      '**/logo.?????*.{webp,avif,svg}', // `@astrojs/image` optimized logo if any
+    ],
+    ignoreURLParametersMatching: [/.*/],
+    runtimeCaching: [{
+      urlPattern: /^\/$/,
+      handler: 'NetworkFirst',
+    }, {
+      urlPattern: settings.logo
+        ? new RegExp(`^${(settings.mini_logo ? `(?:${settings.mini_logo}|${settings.logo})` : settings.logo)}$`)
+        : /^\/(?:img\/uploads\/|img\/)?logo\.(?:png|jpg|jpeg|webp|avif|svg)$/,
+      handler: 'StaleWhileRevalidate',
+      options: {
+        cacheName: 'logo',
+      },
+    }, {
+      urlPattern: /\/((?!(?:admin|assets|img)(\/|$))[^.]+)(\.(?!js|css|xml|txt|png|jpg|jpeg|webp|avif|svg|gif)[^.]+)*$/,
+      handler: 'NetworkFirst',
+      options: {
+        cacheName: 'pages',
+        expiration: {
+          maxEntries: 50,
+          purgeOnQuotaError: true,
+        },
+      },
+    }, {
+      urlPattern: /^\/assets\//,
+      handler: 'StaleWhileRevalidate',
+      options: {
+        cacheName: 'assets',
+      },
+    }, {
+      urlPattern: /^\/img\/uploads\/.*\.(?:png|jpg|jpeg|webp|avif|svg|gif)$/,
+      handler: 'StaleWhileRevalidate',
+      options: {
+        cacheName: 'cms-images',
+        expiration: {
+          maxEntries: 20,
+          maxAgeSeconds: 60 * 60 * 24 * 30, // 30 days
+          purgeOnQuotaError: true,
+        },
+      },
+    }, {
+      urlPattern: /^https:\/\/ecomplus\.io/,
+      handler: 'NetworkFirst',
+      options: {
+        cacheName: 'store-api',
+        expiration: {
+          maxEntries: 50,
+          purgeOnQuotaError: true,
+        },
+      },
+    }, {
+      urlPattern: /^https:\/\/(((\w+\.)?ecoms\d)|(ecom[\w-]+(\.\w+)*\.digitaloceanspaces))\.com.*\/imgs\/normal\//,
+      handler: 'CacheFirst',
+      options: {
+        cacheName: 'product-thumbnails',
+        expiration: {
+          maxEntries: 100,
+          maxAgeSeconds: 60 * 60 * 24 * 30, // 30 days
+          purgeOnQuotaError: true,
+        },
+      },
+    }, {
+      urlPattern: /^https:\/\/(((\w+\.)?ecoms\d)|(ecom[\w-]+(\.\w+)*\.digitaloceanspaces))\.com.*\/imgs\/big\//,
+      handler: 'CacheFirst',
+      options: {
+        cacheName: 'product-pictures',
+        expiration: {
+          maxEntries: 10,
+          maxAgeSeconds: 60 * 60 * 24 * 7, // 7 days
+          purgeOnQuotaError: true,
+        },
+      },
+    }],
+  },
+};
+
+const genAstroConfig = ({
+  site = `https://${domain}`,
+  vitePWAOptions = _vitePWAOptions,
+} = {}) => ({
   output: 'server',
   adapter: node(),
   integrations: [
@@ -22,6 +134,21 @@ export default defineConfig({
     partytown(),
     prefetch(),
     sitemap(),
+    UnoCSS(),
   ],
-  site: `https://${getConfig().domain}`,
+  site,
+  vite: {
+    plugins: [
+      VitePWA(vitePWAOptions),
+    ],
+  },
 });
+
+const astroConfig = genAstroConfig();
+
+// https://astro.build/config
+export default defineConfig(astroConfig);
+
+export { genAstroConfig, astroConfig };
+
+export const vitePWAOptions = _vitePWAOptions;