+++ packages/storefront/scripts/prepare-monorepo.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+rm -rf content && ln -s ../../store/functions/ssr/content content
+rm -rf public && ln -s ../../store/functions/ssr/public public