+++ packages/storefront/scripts/prerelease.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+rm content && cp -r ../../store/functions/ssr/content ./
+rm public && cp -r ../../store/functions/ssr/public ./