+++ packages/storefront/src/env.d.ts
@@ -1,1 +1,3 @@
 /// <reference types="astro/client" />
+/// <reference types="vite-plugin-pwa/client" />
+/// <reference types="../client" />